Fix BUG

+++ optimize_gemm/MMult_1x4x6.h
@@ -6,11 +6,11 @@
 
 /* Routine for computing C = A * B + C */
 
-void AddDot1x4( int, double *, int,  double *, int, double *, int )
+void AddDot1x4( int, float *, int,  float *, int, float *, int )
 
-void MY_MMult( int m, int n, int k, double *a, int lda, 
-                                    double *b, int ldb,
-                                    double *c, int ldc )
+void MY_MMult( int m, int n, int k, float *a, int lda, 
+                                    float *b, int ldb,
+                                    float *c, int ldc )
 {
   int i, j;
 
@@ -25,7 +25,7 @@ void MY_MMult( int m, int n, int k, double *a, int lda,
 }
 
 
-void AddDot1x4( int k, double *a, int lda,  double *b, int ldb, double *c, int ldc )
+void AddDot1x4( int k, float *a, int lda,  float *b, int ldb, float *c, int ldc )
 {
   /* So, this routine computes four elements of C: 
            C( 0, 0 ), C( 0, 1 ), C( 0, 2 ), C( 0, 3 ).  
@@ -37,7 +37,7 @@ void AddDot1x4( int k, double *a, int lda,  double *b, int ldb, double *c, int l
      In this version, we accumulate in registers and put A( 0, p ) in a register */
 
   int p;
-  register double 
+  register float 
     /* hold contributions to
        C( 0, 0 ), C( 0, 1 ), C( 0, 2 ), C( 0, 3 ) */
        c_00_reg,   c_01_reg,   c_02_reg,   c_03_reg,
